+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.0
 - feat: pass kwargs from PageChecker call to httpx request
+- feat: add property `previous_response` on PageChecker
 - chore: update dependencies
 
 ## 0.2.0